Update settings.example.txt and eg001

+++ examples/Admin/eg001CreateNewUserWithActiveStatus.sh
@@ -24,6 +24,25 @@ declare -a Headers=('--header' "Authorization: Bearer ${ACCESS_TOKEN}"
 # Step 3. Construct the request body
 # Create a temporary file to store the JSON body
 request_data=$(mktemp /tmp/request-cw-001.XXXXXX)
+response=$(mktemp /tmp/response-oa.XXXXXX)
+
+# Get user's data
+curl --request GET ${base_path}/v2/organizations/${ORGANIZATION_ID}/users/profile?email=${SIGNER_EMAIL} \
+"${Headers[@]}" \
+--output ${response}
+
+# If the status code returned a response greater than 201, display an error message
+if [[ "$Status" -gt "201" ]]; then
+    echo ""
+    echo "Can't get user's data..."
+    echo ""
+    cat $response
+    exit 0
+fi
+
+# Get group ID and permission profile ID
+group_id=`cat $response | sed 's/.*\"groups\"://' | sed 's/},/\n/g' | sed 's/.*\"id\"://' | sed 's/\".*//g' | sed 's/,//g' | sed -n 2p`
+permission_profile_id=`cat $response | sed 's/.*\"permission_profile\"://' | sed 's/},/\n/g' | sed -n 1p |  sed 's/.*\"id\"://' | sed 's/\".*//g' | sed 's/,//g'`
 
 printf \
 '{
@@ -31,16 +50,16 @@ printf \
   "first_name": "Example",
   "last_name": "Name",
   "email": "examplename42@orobia.net",
-  "auto_activate_memberships": true,
+  "auto_activate_memberships": false,
   "accounts": [
     {
-      "id": ${ACCOUNT_ID},
+      "id": \"'${ACCOUNT_ID}'\",
       "permission_profile": {
-        "id": xxxxxxx,
+        "id": \"'${permission_profile_id}'\",
       },
       "groups": [
         {
-          "id": xxxxxxx,
+          "id": \"'${group_id}'\",
         }
       ]
     }
@@ -49,24 +68,32 @@ printf \
 ' >>$request_data
 
 # Call the DocuSign Admin API
-response=$(mktemp /tmp/response-oa.XXXXXX)
+response2=$(mktemp /tmp/response-oa.XXXXXX)
 curl --request POST ${base_path}/v2/organizations/${ORGANIZATION_ID}/users \
 "${Headers[@]}" \
 --data-binary @${request_data} \
---output ${response}
+--output ${response2}
+
+# If the status code returned a response greater than 201, display an error message
+if [[ "$Status" -gt "201" ]]; then
+    echo ""
+    echo "Failed to create a new user"
+    echo ""
+    cat $response
+    exit 0
+fi
 
 echo ''
 echo 'Response:'
 echo ''
-cat $response
+cat $response2
 echo ''
 
 # Remove the temporary files
 rm "$request_data"
 rm "$response"
+rm "$response2"
 
 echo ""
 echo "Done."
 echo ""
-
-echo "eg001CreateNewUserWithActiveStatus.sh - Work in progress..."
